Fix build, add impls of missing compiler builtins
On platforms that don't have native 32-bit shift instructions, LLVM has
to lower these instructions into platform-specific subroutines that can
handle a shift of a 32-bit type using the native hardware registers. In
the case of AVR, only 8-bit registers are available, and shifts of 16
bit types are implemented by rust's compiler_builtins crate. Shifts of
32 bit types were not implemented at the time of writing (they should be
   landing soon), so I implemented them by depending on existing support
of shifting 16 bit types in compiler_builtins for now.

libgcc also provides some of these compiler "lowering" routines for
different platforms (with the same symbol names), but since rust is
compiled with LLVM it might make more sense to link in compiler-rt.
Unclear, because we also link in avr-libc, which seems to depend on
libgcc. compiler-rt is being replaced by rust's compiler_builtins, so we
should be able to get all the symbols rust needs without compiler-rt.

Unlinking libgcc should maybe break avr-libc but, not any rust code. If
rust code is calling symbols from libgcc that probably means that they
need to be implemented in compiler_builtins. `__udivmodhi4` seems to be
a prime candidate for this.